refactor(toolkit): move context providers
