+++ packages/@aws-cdk/toolkit-lib/lib/api/aws-cdk.ts
@@ -2,6 +2,3 @@
 
 // APIs
 export { CloudWatchLogEventMonitor, findCloudWatchLogGroups } from '../../../../aws-cdk/lib/api/logs-monitor';
-
-// Context Providers
-export * as contextproviders from '../../../../aws-cdk/lib/context-providers';
