+++ packages/@aws-cdk/toolkit-lib/lib/api/shared-private.ts
@@ -4,3 +4,6 @@ export * from '../../../tmp-toolkit-helpers/src/api/io/private';
 export * from '../../../tmp-toolkit-helpers/src/private';
 export * from '../../../tmp-toolkit-helpers/src/api';
 export * as cfnApi from '../../../tmp-toolkit-helpers/src/api/deployments/cfn-api';
+
+// Context Providers
+export * as contextproviders from '../../../tmp-toolkit-helpers/src/context-providers';
